requirements.txt
Click==7.0
Flask==1.0.2
Flask-Dropzone==1.5.3
Flask-Uploads==0.2.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.4.1
Werkzeug==0.14.1